+++ src/Plugins/Hawkeye.DecompilePlugin/Properties/AssemblyInfo.cs
@@ -1,0 +1,17 @@
+﻿using System.Reflection;
+
+[assembly: AssemblyTitle("Hawkeye.DecompilePlugin")]
+[assembly: AssemblyDescription("Decompilation plugins for Hawkeye 2")]
+[assembly: AssemblyConfiguration("")]
+[assembly: AssemblyCompany("https://github.com/odalet/Hawkeye2")]
+[assembly: AssemblyProduct("Hawkeye 2")]
+[assembly: AssemblyCopyright("")]
+[assembly: AssemblyTrademark("Hawkeye")]
+[assembly: AssemblyCulture("")]
+[assembly: AssemblyVersion(AssemblyVersionInfo.Version)]
+[assembly: AssemblyFileVersion(AssemblyVersionInfo.Version)]
+
+internal static class AssemblyVersionInfo
+{
+    public const string Version = "1.0.0.0";
+}